Refactor 1: file structure
