+++ packages/govern-console/package.json
@@ -19,7 +19,7 @@
     "apollo-link-http": "^1.5.17",
     "bn.js": "^5.1.3",
     "clipboard-polyfill": "^2.8.6",
-    "dayjs": "^1.8.35",
+    "date-fns": "^2.16.1",
     "ethers": "^5.0.14",
     "graphql": "^15.0.0",
     "graphql-tag": "^2.10.3",
