+++ packages/govern-console/src/lib/known-roles.ts
@@ -1,0 +1,10 @@
+export const KNOWN_QUEUE_ROLES = new Map([
+  ['0x977d8964', 'schedule'],
+  ['0x3a139c71', 'execute'],
+  ['0x70576158', 'challenge'],
+  ['0x72896761', 'configure'],
+  ['0xc04c87b8', 'veto'],
+  ['0x586df604', 'ROOT_ROLE'],
+])
+
+export const KNOWN_GOVERN_ROLES = new Map([['0xc2d85afc', 'exec']])
