+++ packages/govern-console/src/lib/utils.ts
@@ -1,0 +1,8 @@
+export function log(...params) {
+  if (
+    process.env.NODE_ENV !== 'production' &&
+    process.env.NODE_ENV !== 'test'
+  ) {
+    console.log(...params)
+  }
+}
